#152 implement permissions

+++ unitrade-backend/src/main/java/ca/mcgill/ecse428/unitrade/unitradebackend/controller/PostRestController.java
@@ -7,8 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -21,7 +19,6 @@ import org.springframework.web.bind.annotation.RestController;
 import ca.mcgill.ecse428.unitrade.unitradebackend.dto.Request.PostRequestDto;
 import ca.mcgill.ecse428.unitrade.unitradebackend.dto.Response.PostResponseDto;
 import ca.mcgill.ecse428.unitrade.unitradebackend.model.Post;
-import ca.mcgill.ecse428.unitrade.unitradebackend.security.CustomUserDetails;
 import ca.mcgill.ecse428.unitrade.unitradebackend.service.PostService;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
@@ -43,16 +40,7 @@ public class PostRestController {
     @PostMapping(value = { "/post" })
     public ResponseEntity<PostResponseDto> createPost(@RequestBody PostRequestDto body) {
 
-        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-        // Get the email of the authenticated user
-        Long authId = null;
-
-        if (principal instanceof UserDetails) {
-            authId = ((CustomUserDetails) principal).getId();
-        } else {
-            return new ResponseEntity<PostResponseDto>(HttpStatus.EXPECTATION_FAILED);
-        }
+        Long authId = ControllerHelper.getAuthenticatedUserId();
 
         Post post = postService.createPost(
                 body.getTitle(),
@@ -88,6 +76,15 @@ public class PostRestController {
     @ResponseStatus(HttpStatus.OK)
     @DeleteMapping(value = { "/post/{id}" })
     public ResponseEntity<PostResponseDto> deletePost(@PathVariable("id") Long id) {
+
+
+        Long authId = ControllerHelper.getAuthenticatedUserId();
+
+        Post post = postService.getPost(id);
+
+        if (post.getPoster().getId() != authId)
+            return new ResponseEntity<PostResponseDto>(HttpStatus.UNAUTHORIZED);
+
         postService.deletePost(id);
         return new ResponseEntity<PostResponseDto>(HttpStatus.OK);
     }
